refactor: Linter happy
